Fix bug that can not handle Chinese chars when generate toc.
Also fix bug that the indent of toc.
Before fix, some section of toc will be show as code setion, and the
generated fragment link won't work when it contains Chinese chars.

We replace all space to '-', and replace all none digital/alphabetic
char to '-', this will let the fragment links work fine in browser.